Solution to Check Triangle Validity

+++ Check Triangle Validity@@6788dd63490637cfc44e1b93/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 #include <math.h>
 int main(){
-    int a,b;
+    float a,b;
     scanf("%d%d",&a,&b);
-    int c=((a*a)+(b*b));
-    int square=sqrt(c);
-    int d=sqrt(c);
+    float c=((a*a)+(b*b));
+    float square=sqrt(c);
+    float d=sqrt(c);
     if(d==c){
         printf("Valid");
     }
